+++ app/response_processor.py
@@ -125,8 +125,9 @@ class ResponseProcessor:
 
             try:
                 self.logger.info("About to publish receipt into rrm queue")
-                self.rrm_publisher.publish_message(dumps(receipt_json),
-                                                   secret=settings.SDX_COLLECT_SECRET)
+                self.rrm_publisher.publish(dumps(receipt_json),
+                                           headers={'tx_id': decrypted_json['tx_id']},
+                                           secret=settings.SDX_COLLECT_SECRET)
             except PublishMessageError as e:
                 self.logger.error("Unsuccesful publish", error=e)
                 raise RetryableError