+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### Unreleased
   - Publish all surveys to same downstream by removing cora specific queueing
   - Remove sdx common clone in Dockerfile
+  - Send headers to RRM receipt publisher
   
 ### 2.0.0 2017-09-11
   - Ensure integrity and version of library dependencies